ファイル構造の変更
