+++ model/util.go
@@ -1,0 +1,23 @@
+package model
+
+import (
+	"fmt"
+	"os"
+	"log"
+
+	"github.com/jmoiron/sqlx"
+)
+
+var (
+	Db *sqlx.DB
+)
+
+
+//Establish データベースに接続
+func Establish(){
+	_db, err := sqlx.Connect("mysql", fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=utf8&parseTime=True&loc=Local", os.Getenv("DB_USERNAME"), os.Getenv("DB_PASSWORD"), os.Getenv("DB_HOSTNAME"), os.Getenv("DB_PORT"), os.Getenv("DB_DATABASE")))
+	if err != nil {
+		log.Fatalf("Cannot Connect to Database: %s", err)
+	}
+	Db = _db
+}
